pyproject.toml
[project]
name = "news-grpo"
version = "0.1.0"
description = "GRPO training platform for news data using VERL"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "supabase>=2.24.0",
    "torch>=2.5.0",
    "transformers>=4.47.0",
    "accelerate>=1.2.0",
    "datasets>=3.2.0",
    "numpy<2.0.0",
    "vllm>=0.6.5",
    "ray>=2.10.0",
    "peft>=0.13.0",
    "bitsandbytes>=0.45.0",
    "pyyaml>=6.0",
    "wandb>=0.16.0",
    "flash-attn>=2.6.0",
    "verl @ git+https://github.com/volcengine/verl.git",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[tool.uv]

[tool.uv.extra-build-dependencies]
flash-attn = ["torch", "packaging", "ninja"]
